add .hpp file and remove some functions to the .hpp file
