+++ tests/2d_examples/test_2d_eulerian_flow_around_cylinder/2d_eulerian_flow_around_cylinder.h
@@ -7,7 +7,7 @@
 #ifndef EULERIAN_FLOW_AROUND_CYLINDER_H
 #define EULERIAN_FLOW_AROUND_CYLINDER_H
 #include "common_shared_eulerian_classes.h" // shared eulerian classes for weakly-compressible and compressible fluid.
-#include "common_weakly_compressible_eulerian_classes.h" // eulerian classes for weakly compressible fluid only.
+#include "common_weakly_compressible_eulerian_classes.hpp" // eulerian classes for weakly compressible fluid only.
 using namespace SPH;
 //----------------------------------------------------------------------
 //	Basic geometry parameters and numerical setup.
